+++ dessia_common/workflow/core.py
@@ -18,6 +18,7 @@ import traceback as tb
 import networkx as nx
 
 import dessia_common.errors
+from dessia_common.graph import get_column_by_node
 from dessia_common.templates import workflow_template
 from dessia_common import DessiaObject, is_sequence, JSONSCHEMA_HEADER, jsonschema_from_annotation, \
     deserialize_argument, set_default_value, prettyname, serialize_dict, DisplaySetting
@@ -43,80 +44,93 @@ class Variable(DessiaObject):
     _eq_is_data_eq = False
     has_default_value: bool = False
 
-    def __init__(self, name: str = ''):
+    def __init__(self, name: str = '', position = None):
         """
         Variable for workflow
         """
         DessiaObject.__init__(self, name=name)
-        self.position = None
+        if position is None:
+            self.position = (0, 0)
+        else:
+            self.position = position
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
         dict_ = DessiaObject.base_dict(self)
-        dict_.update({'has_default_value': self.has_default_value})
+        dict_.update({'has_default_value': self.has_default_value,
+                      'position': self.position})
         return dict_
+
+
+    def _to_script(self) -> ToScriptElement:
+        script = self._get_to_script_elements()
+        script.declaration = f"{self.__class__.__name__}({script.declaration})"
+
+        script.imports.append(self.full_classname)
+        return script
+
+    def _get_to_script_elements(self):
+        declaration = f"name='{self.name}', position={self.position}"
+        return ToScriptElement(declaration=declaration, imports=[], imports_as_is=[])
 
 
 class TypedVariable(Variable):
     has_default_value: bool = False
 
-    def __init__(self, type_: Type, name: str = ''):
+    def __init__(self, type_: Type, name: str = '', position = None):
         """
         Variable for workflow with a typing
         """
-        Variable.__init__(self, name=name)
+        Variable.__init__(self, name=name, position=position)
         self.type_ = type_
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
-        dict_ = DessiaObject.base_dict(self)
-        dict_.update({'type_': serialize_typing(self.type_), 'has_default_value': self.has_default_value})
+        dict_ = super().to_dict(use_pointers, memo, path)
+        dict_.update({'type_': serialize_typing(self.type_)})
         return dict_
 
     @classmethod
     def dict_to_object(cls, dict_: JsonSerializable, force_generic: bool = False,
                        global_dict=None, pointers_memo: Dict[str, Any] = None, path: str = '#') -> 'TypedVariable':
         type_ = deserialize_typing(dict_['type_'])
-        return cls(type_=type_, name=dict_['name'])
+        return cls(type_=type_, name=dict_['name'], position=dict_.get("position"))
 
     def copy(self, deep: bool = False, memo=None):
         return TypedVariable(type_=self.type_, name=self.name)
 
-    def _to_script(self) -> ToScriptElement:
-        script = f"TypedVariable(type_={serialize_typing(self.type_)}, name='{self.name}')\n"
+    def _get_to_script_elements(self) -> ToScriptElement:
+        script = super()._get_to_script_elements()
 
-        imports = [self.full_classname]
-        imports_as_is = None
-        if "builtins" in serialize_typing(self.type_):
-            imports_as_is = ["builtins"]
-        else:
-            imports.append(serialize_typing(self.type_))
-        return ToScriptElement(declaration=script, imports=imports, imports_as_is=imports_as_is)
+        script.declaration += f", type_={self.type_.__name__}"
+
+        if "builtins" not in serialize_typing(self.type_):
+            script.imports.append(serialize_typing(self.type_))
+        return script
 
 
 class VariableWithDefaultValue(Variable):
     has_default_value: bool = True
 
-    def __init__(self, default_value: Any, name: str = ''):
+    def __init__(self, default_value: Any, name: str = '', position=None):
         """
         A variable with a default value
         """
-        Variable.__init__(self, name=name)
+        Variable.__init__(self, name=name, position=position)
         self.default_value = default_value
 
 
 class TypedVariableWithDefaultValue(TypedVariable):
     has_default_value: bool = True
 
-    def __init__(self, type_: Type, default_value: Any, name: str = ''):
+    def __init__(self, type_: Type, default_value: Any, name: str = '', position = None):
         """
         Workflow variables wit a type and a default value
         """
-        TypedVariable.__init__(self, type_=type_, name=name)
+        TypedVariable.__init__(self, type_=type_, name=name, position=position)
         self.default_value = default_value
 
     def to_dict(self, use_pointers: bool = True, memo=None, path: str = '#'):
-        dict_ = DessiaObject.base_dict(self)
-        dict_.update({'type_': serialize_typing(self.type_), 'default_value': serialize(self.default_value),
-                      'has_default_value': self.has_default_value})
+        dict_ = super().to_dict(use_pointers, memo, path)
+        dict_.update({'default_value': serialize(self.default_value)})
         return dict_
 
     @classmethod
@@ -124,7 +138,7 @@ class TypedVariableWithDefaultValue(TypedVariable):
                        pointers_memo: Dict[str, Any] = None, path: str = '#') -> 'TypedVariableWithDefaultValue':
         type_ = deserialize_typing(dict_['type_'])
         default_value = deserialize(dict_['default_value'], global_dict=global_dict, pointers_memo=pointers_memo)
-        return cls(type_=type_, default_value=default_value, name=dict_['name'])
+        return cls(type_=type_, default_value=default_value, name=dict_['name'], position=dict_.get('position'))
 
     def copy(self, deep: bool = False, memo=None):
         """
@@ -138,6 +152,12 @@ class TypedVariableWithDefaultValue(TypedVariable):
             memo = {}
         copied_default_value = deepcopy_value(self.default_value, memo=memo)
         return TypedVariableWithDefaultValue(type_=self.type_, default_value=copied_default_value, name=self.name)
+
+    def _to_script(self) -> ToScriptElement:
+        warnings.warn("to_script method is not implemented for TypedVariableWithDefaultValue yet. "
+                      "We are losing the default value as we call the TypedVariable method")
+        casted_variable = TypedVariable(type_=self.type_, name=self.name, position=self.position)
+        return casted_variable._to_script()
 
 
 NAME_VARIABLE = TypedVariable(type_=str, name="Result Name")
@@ -163,13 +183,16 @@ class Block(DessiaObject):
     _non_serializable_attributes = []
 
     def __init__(self, inputs: List[Variable], outputs: List[Variable],
-                 position: Tuple[float, float] = (0, 0), name: str = ''):
+                 position: Tuple[float, float] = None, name: str = ''):
         """
         An Abstract block. Do not instantiate alone
         """
+        if position is None:
+            self.position = (0, 0)
+        else:
+            self.position = position
         self.inputs = inputs
         self.outputs = outputs
-        self.position = position
         DessiaObject.__init__(self, name=name)
 
     def equivalent_hash(self):
@@ -210,6 +233,9 @@ class Block(DessiaObject):
         """
         block_docstring = {i: EMPTY_PARSED_ATTRIBUTE for i in self.inputs}
         return block_docstring
+
+    def base_script(self) -> str:
+        return f"name='{self.name}', position={self.position}"
 
 
 class Pipe(DessiaObject):
@@ -299,6 +325,10 @@ class Workflow(Block):
                      'default_value': '', 'python_typing': 'builtins.str'}
         }
     }
+
+    @property
+    def nodes(self):
+        return self.blocks + self.nonblock_variables
 
     def __init__(self, blocks, pipes, output, *, imposed_variable_values=None,
                  detached_variables: List[TypedVariable] = None, description: str = "",
@@ -663,7 +693,7 @@ class Workflow(Block):
         if memo is None:
             memo = {}
 
-        self.refresh_blocks_positions()
+        # self.refresh_blocks_positions()
         dict_ = Block.to_dict(self)
         dict_['object_class'] = 'dessia_common.workflow.core.Workflow'  # TO force migrating from dessia_common.workflow
         blocks = [b.to_dict() for b in self.blocks]
@@ -1119,51 +1149,42 @@ class Workflow(Block):
             return False
         return True
 
-    def layout(self, min_horizontal_spacing=300, min_vertical_spacing=200, max_height=800, max_length=1500):
-        """
-        Computes workflow layout
-        """
-        coordinates = {}
-        elements_by_distance = {}
-        if self.output:
-            for element in self.blocks + self.nonblock_variables:
-                distances = []
-                paths = nx.all_simple_paths(self.graph, element, self.output)
-                for path in paths:
-                    distance = 1
-                    for path_element in path[1:-1]:
-                        if path_element in self.blocks + self.nonblock_variables:
-                            distance += 1
-                    distances.append(distance)
-                try:
-                    distance = max(distances)
-                except ValueError:
-                    distance = 3
-                if distance in elements_by_distance:
-                    elements_by_distance[distance].append(element)
-                else:
-                    elements_by_distance[distance] = [element]
+    @property
+    def layout_graph(self) -> nx.DiGraph:
+        graph = nx.DiGraph()
+        graph.add_nodes_from(self.nodes) #does not handle detached_variable
 
-        if len(elements_by_distance) != 0:
-            max_distance = max(elements_by_distance.keys())
-        else:
-            max_distance = 3  # TODO: this is an awfull quick fix
+        for pipe in self.pipes:
+            if pipe.input_variable in self.nonblock_variables:
+                input_node = pipe.input_variable
+            else:
+                input_node = self.block_from_variable(pipe.input_variable)
+            output_block = self.block_from_variable(pipe.output_variable)
+            graph.add_edge(input_node, output_block)
 
-        horizontal_spacing = max(min_horizontal_spacing, max_length / max_distance)
+        return graph
 
-        for i, distance in enumerate(sorted(elements_by_distance.keys())[::-1]):
-            vertical_spacing = min(min_vertical_spacing, max_height / len(elements_by_distance[distance]))
-            for j, element in enumerate(elements_by_distance[distance]):
-                coordinates[element] = (i * horizontal_spacing, (j + 0.5) * vertical_spacing)
-        return coordinates
-
-    def refresh_blocks_positions(self):
+    def graph_columns(self, graph):
         """
-        Recomputes block positions
+        :returns: list[ColumnLayout] where ColumnLayout is list[node_index]
         """
-        coordinates = self.layout()
-        for block, coordinate in coordinates.items():
-            block.position = coordinate
+        column_by_node = get_column_by_node(graph)
+        nodes_by_column = {}
+        for node, column_index in column_by_node.items():
+            node_index = self.nodes.index(node)
+            nodes_by_column[column_index] = nodes_by_column.get(column_index, []) + [node_index]
+
+        return [column_list for column_list in nodes_by_column.values()]
+
+    def layout(self):
+        """
+        :returns: list[GraphLayout] where GraphLayout is list[ColumnLayout] and ColumnLayout is list[node_index]
+        """
+        digraph = self.layout_graph
+        graph = digraph.to_undirected()
+        connected_components = nx.connected_components(graph)
+
+        return [self.graph_columns(digraph.subgraph(cc)) for cc in list(connected_components)]
 
     def plot_graph(self):
         """
@@ -1221,11 +1242,49 @@ class Workflow(Block):
         """
         return WorkflowState(self, input_values=input_values, name=name)
 
+    def jointjs_layout(self, min_horizontal_spacing=300, min_vertical_spacing=200, max_height=800, max_length=1500):
+        """
+        Deprecated workflow layout. Used only in jointjs_data method.
+        """
+        coordinates = {}
+        elements_by_distance = {}
+        if self.output:
+            for element in self.nodes:
+                distances = []
+                paths = nx.all_simple_paths(self.graph, element, self.output)
+                for path in paths:
+                    distance = 1
+                    for path_element in path[1:-1]:
+                        if path_element in self.blocks + self.nonblock_variables:
+                            distance += 1
+                    distances.append(distance)
+                try:
+                    distance = max(distances)
+                except ValueError:
+                    distance = 3
+                if distance in elements_by_distance:
+                    elements_by_distance[distance].append(element)
+                else:
+                    elements_by_distance[distance] = [element]
+
+        if len(elements_by_distance) != 0:
+            max_distance = max(elements_by_distance.keys())
+        else:
+            max_distance = 3  # TODO: this is an awfull quick fix
+
+        horizontal_spacing = max(min_horizontal_spacing, max_length / max_distance)
+
+        for i, distance in enumerate(sorted(elements_by_distance.keys())[::-1]):
+            vertical_spacing = min(min_vertical_spacing, max_height / len(elements_by_distance[distance]))
+            for j, element in enumerate(elements_by_distance[distance]):
+                coordinates[element] = (i * horizontal_spacing, (j + 0.5) * vertical_spacing)
+        return coordinates
+
     def jointjs_data(self):
         """
         Computes the data needed for jointjs ploting
         """
-        coordinates = self.layout()
+        coordinates = self.jointjs_layout()
         blocks = []
         for block in self.blocks:
             # TOCHECK Is it necessary to add is_workflow_input/output for outputs/inputs ??
@@ -1331,50 +1390,48 @@ class Workflow(Block):
             raise ValueError("A workflow output must be set")
 
         # --- Blocks ---
-        blockstr = ""
+        blocks_str = ""
         imports = []
         imports_as_is = []
         for iblock, block in enumerate(self.blocks):
             block_script = block._to_script()
             imports.extend(block_script.imports)
             if block_script.before_declaration is not None:
-                blockstr += f"{block_script.before_declaration}\n"
-            blockstr += f'{prefix}block_{iblock} = {block_script.declaration}\n'
-        blockstr += f"{prefix}blocks = [{', '.join([prefix + 'block_' + str(i) for i in range(len(self.blocks))])}]\n"
+                blocks_str += f"{block_script.before_declaration}\n"
+            blocks_str += f'{prefix}block_{iblock} = {block_script.declaration}\n'
+        blocks_str += f"{prefix}blocks = [{', '.join([prefix + 'block_' + str(i) for i in range(len(self.blocks))])}]\n"
+
+        # --- NBVs ---
+        nbvs_str = ""
+        for nbv_index, nbv in enumerate(self.nonblock_variables):
+            nbv_script = nbv._to_script()
+            imports.extend(nbv_script.imports)
+            imports_as_is.extend(nbv_script.imports_as_is)
+            nbvs_str += f"{prefix}variable_{nbv_index} = {nbv_script.declaration}\n"
 
         # --- Pipes ---
-        pipestr = ""
-        variable_index = 0
+        pipes_str = ""
         for ipipe, pipe in enumerate(self.pipes):
             input_index = self.variable_indices(pipe.input_variable)
             if isinstance(input_index, int):  # NBV handling
-                input_name = f'{prefix}variable_{variable_index}'
-                input_script_elements = pipe.input_variable._to_script()
-                imports.extend(input_script_elements.imports)
-                imports_as_is.extend((input_script_elements.imports_as_is))
-                pipestr += f'{input_name} = {input_script_elements.declaration}'
-                variable_index += 1
+                input_name = f'{prefix}variable_{input_index}'
             else:
                 input_name = f"{prefix}block_{input_index[0]}.outputs[{input_index[2]}]"
 
             output_index = self.variable_indices(pipe.output_variable)
             if isinstance(output_index, int):  # NBV handling
-                output_name = f'{prefix}variable_{variable_index}'
-                output_script_elements = pipe.output_variable._to_script()
-                imports.extend(output_script_elements.imports)
-                imports_as_is.extend(output_script_elements.imports_as_is)
-                pipestr += f'{output_name } = {output_script_elements.declaration}'
-                variable_index += 1
+                output_name = f'{prefix}variable_{output_index}'
             else:
                 output_name = f"{prefix}block_{output_index[0]}.inputs[{output_index[2]}]"
-            pipestr += f"{prefix}pipe_{ipipe} = Pipe({input_name}, {output_name})\n"
-        pipestr += f"{prefix}pipes = [{', '.join([prefix + 'pipe_' + str(i) for i in range(len(self.pipes))])}]\n"
+            pipes_str += f"{prefix}pipe_{ipipe} = Pipe({input_name}, {output_name})\n"
+        pipes_str += f"{prefix}pipes = [{', '.join([prefix + 'pipe_' + str(i) for i in range(len(self.pipes))])}]\n"
 
         # --- Building script ---
         output_name = f"{prefix}block_{workflow_output_index[0]}.outputs[{workflow_output_index[2]}]"
 
-        full_script = f"{blockstr}\n" \
-                      f"{pipestr}\n" \
+        full_script = f"{blocks_str}\n" \
+                      f"{nbvs_str}\n" \
+                      f"{pipes_str}\n" \
                       f"{prefix}workflow = " \
                       f"Workflow({prefix}blocks, {prefix}pipes, output={output_name}, name='{self.name}')\n"