+++ CHANGELOG.md
@@ -10,12 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Sampler class
 - (De)Serialization handles 'typing.Type'
+- Workflow: handle position in to_dict / dict_to_object process
 
 ### Breaking Changes
 
 
 ### Changes
-
+- Workflow: improve layout method
 
 ### Performance
 - switch is_jsonable to orjson